Clean up some stuff
Formatting, mostly. I'm very specific about these things. Anyway, I also
want to go through and change everything so that it will actually work
to test with. I want to leave the master branch alone so that we still
have all the CAN stuff, but I want this to be much closer to a real
robot so that we can do some better testing. I want to write a new
TestingDrivetrain class and have every old reference changed to fit the
new one so that testing will more closely mirror production.