Play with potentiometer
Start writing some stuff to read from it, but it's not working, largely
because the potentiometer itself isn't working right. I don't know if
potentiometers just act like this or if this one is wrong, but it is
acting really werid. Also, I screwed up a constant, so the right side is
now controlled by the trigger. I should fix that.