Get master/slave to work
When we put all four motors into the constructor, I think that their
assignments may have been overwritten and they thought that they were
masters. I set it up so that only the actual masters go into the
RobotDrive, and it works. Also, it assumes that the driven motors are in
the back, although this is nowhere in the documentation. Go figure.