Fix issue with PID?
Maybe? It has been theorized that we were calling the wrong RobotDrive,
one whose maxSpeed wasn't set, so it wasn't actually giving it hardly
any power. talon.set() was expecting -256-256 or something, and all it
was getting was -1-1, in other words. It followed that, if we remove the
inheritance structure that used to exist, and just had everything done
manually, it would be fine. We'll see how that goes. I'm going to test
this now.