Reimplement CANDrivetrainSubsystem
The PID constants seem wrong, but nothing seems to change when I change
them. So that's nice. I changed them to the ones from time ago, but I
didn't get to test it. I'll fiddle with it later. We're about to get to
test with a real robot, though, so that'll be fun.