Add some autonomous code
There is now a command that should (*should*) move the robot forward
some amount of feet, and another one that should move it for a certain
amount of time. We should also maybe have a command that turns the robot
a certain amount of degrees, but we might already have that. I'm going
to look through the stuff we already have and see what's going on.