Write some new fancy drive code
I thought that it would be cool if the computer would sum all the inputs
so that it would control more naturally. For example, if you are
pressing the right trigger and also pushing the right stick up, it will
turn left. So I wrote some code for that. Who knows how well it will
work, but I thought I should try it.